woohoo: ProductTest passing, fully autocompiled. new class arb.expressions.nodes.Variable at pos 2 new class arb.expressions.nodes.nary.Product at pos 2 new class arb.expressions.nodes.Variable at pos 4 new class arb.expressions.nodes.LiteralConstant at pos 6 new class arb.expressions.nodes.LiteralConstant at pos 8 Generating \prod_{k = 1}^{3}{k} new class arb.expressions.nodes.Variable at pos -1 new class arb.expressions.nodes.Variable at pos 3 Generating Variable[reference=INPUT(k), type=arb.Integer] Declaring variable of factorℝ1: k=0 // class version 65.65535 (-65471) // access flags 0x21 // signature Ljava/lang/Object;Larb/functions/Function<Larb/Integer;Larb/Real;>; // declaration: factorℝ1 implements arb.functions.Function<arb.Integer, arb.Real> public class factorℝ1 implements arb/functions/Function {
  // access flags 0x1
  // signature (Larb/Integer;IILarb/Real;)Larb/Real;
  // declaration: arb.Real evaluate(arb.Integer, int, int, arb.Real)
  public evaluate(Ljava/lang/Object;IILjava/lang/Object;)Ljava/lang/Object;
   L0
    ALOAD 1
    CHECKCAST arb/Integer
    ALOAD 4
    CHECKCAST arb/Real
    SWAP
    INVOKEVIRTUAL arb/Real.set (Larb/Integer;)Larb/Real;
    ARETURN
   L1
    LOCALVARIABLE this Lfactorℝ1; L0 L1 0
    LOCALVARIABLE in Larb/Integer; L0 L1 1
    LOCALVARIABLE order I L0 L1 2
    LOCALVARIABLE bits I L0 L1 3
    LOCALVARIABLE result Larb/Real; L0 L1 4
    MAXSTACK = 10
    MAXLOCALS = 10

  // access flags 0x1
  public Larb/Integer; k

  // access flags 0x1
  public <init>()V
    ALOAD 0
    INVOKESPECIAL java/lang/Object.<init> ()V
    RETURN
    MAXSTACK = 10
    MAXLOCALS = 10
}
Loading factorℝ1
defining factorℝ1
registerFunction( functionName = factorℝ1, function = factorℝ1@5ef60048,
domainType=class arb.Integer, rangeType=class java.lang.Object )
POP
getField(fieldName=k, fieldTypeSignature=Larb/Integer;
loadThis ALOAD 0
POP
loadThis ALOAD 0
loadThis ALOAD 0
loadThis ALOAD 0
loadThis ALOAD 0
POP
getField(fieldName=k, fieldTypeSignature=Larb/Integer;
loadThis ALOAD 0
loadThis ALOAD 0
Declaring variable of _k_k_1_3_: k=0
// class version 65.65535 (-65471)
// access flags 0x21
public class _k_k_1_3_ implements arb/functions/real/RealNullaryFunction
{

  // access flags 0x1
  // signature (Ljava/lang/Void;IILarb/Real;)Larb/Real;
  // declaration: arb.Real evaluate(java.lang.Void, int, int, arb.Real)
  public evaluate(Ljava/lang/Object;IILjava/lang/Object;)Ljava/lang/Object;
   L0
    ALOAD 0
    GETFIELD _k_k_1_3_.productℝ1 : Larb/Real;
    INVOKEVIRTUAL arb/Real.multiplicativeIdentity ()Larb/Real;
    POP
    ALOAD 0
    GETFIELD _k_k_1_3_.k : Larb/Integer;
    ALOAD 0
    GETFIELD _k_k_1_3_.startIndex : Larb/Integer;
    INVOKEVIRTUAL arb/Integer.set (Larb/Integer;)Larb/Integer;
    POP
   L1
   FRAME SAME
    ALOAD 0
    GETFIELD _k_k_1_3_.productℝ1 : Larb/Real;
    ALOAD 0
    GETFIELD _k_k_1_3_.factorℝ1 : Lfactorℝ1;
    ALOAD 0
    GETFIELD _k_k_1_3_.k : Larb/Integer;
    ILOAD 3
    ALOAD 0
    GETFIELD _k_k_1_3_.factorValueℝ1 : Larb/Real;
    INVOKEINTERFACE arb/functions/Function.evaluate
(Ljava/lang/Object;ILjava/lang/Object;)Ljava/lang/Object; (itf)
    CHECKCAST arb/Real
    ILOAD 3
    INVOKEVIRTUAL arb/Real.mul (Larb/Real;I)Larb/Real;
    POP
    ALOAD 0
    GETFIELD _k_k_1_3_.k : Larb/Integer;
    INVOKEVIRTUAL arb/Integer.increment ()Larb/Integer;
    ALOAD 0
    GETFIELD _k_k_1_3_.endIndex : Larb/Integer;
    INVOKEVIRTUAL arb/Integer.compareTo (Larb/Integer;)I
    IFLE L1
    ALOAD 4
    CHECKCAST arb/Real
    ALOAD 0
    GETFIELD _k_k_1_3_.productℝ1 : Larb/Real;
    INVOKEVIRTUAL arb/Real.set (Larb/Real;)Larb/Real;
    ARETURN
   L2
    LOCALVARIABLE this L_k_k_1_3_; L0 L2 0
    LOCALVARIABLE in Ljava/lang/Void; L0 L2 1
    LOCALVARIABLE order I L0 L2 2
    LOCALVARIABLE bits I L0 L2 3
    LOCALVARIABLE result Larb/Real; L0 L2 4
    MAXSTACK = 10
    MAXLOCALS = 10

  // access flags 0x0
  Larb/Integer; startIndex

  // access flags 0x0
  Larb/Integer; endIndex

  // access flags 0x1
  public Larb/Integer; k

  // access flags 0x1
  public Larb/Real; factorValueℝ1

  // access flags 0x1
  public Larb/Real; productℝ1

  // access flags 0x11
  public final Lfactorℝ1; factorℝ1

  // access flags 0x1
  public <init>()V
    ALOAD 0
    INVOKESPECIAL java/lang/Object.<init> ()V
    ALOAD 0
    NEW factorℝ1
    DUP
    INVOKESPECIAL factorℝ1.<init> ()V
    PUTFIELD _k_k_1_3_.factorℝ1 : Lfactorℝ1;
    ALOAD 0
    NEW arb/Integer
    DUP
    LDC "1"
    INVOKESPECIAL arb/Integer.<init> (Ljava/lang/String;)V
    PUTFIELD _k_k_1_3_.startIndex : Larb/Integer;
    ALOAD 0
    NEW arb/Integer
    DUP
    LDC "3"
    INVOKESPECIAL arb/Integer.<init> (Ljava/lang/String;)V
    PUTFIELD _k_k_1_3_.endIndex : Larb/Integer;
    ALOAD 0
    NEW arb/Real
    DUP
    INVOKESPECIAL arb/Real.<init> ()V
    PUTFIELD _k_k_1_3_.factorValueℝ1 : Larb/Real;
    ALOAD 0
    NEW arb/Real
    DUP
    INVOKESPECIAL arb/Real.<init> ()V
    PUTFIELD _k_k_1_3_.productℝ1 : Larb/Real;
    RETURN
    MAXSTACK = 10
    MAXLOCALS = 10

  // access flags 0x1
  public close()V
    ALOAD 0
    GETFIELD _k_k_1_3_.startIndex : Larb/Integer;
    INVOKEVIRTUAL arb/Integer.close ()V
    ALOAD 0
    GETFIELD _k_k_1_3_.endIndex : Larb/Integer;
    INVOKEVIRTUAL arb/Integer.close ()V
    ALOAD 0
    GETFIELD _k_k_1_3_.factorValueℝ1 : Larb/Real;
    INVOKEVIRTUAL arb/Real.close ()V
    ALOAD 0
    GETFIELD _k_k_1_3_.productℝ1 : Larb/Real;
    INVOKEVIRTUAL arb/Real.close ()V
    RETURN
    MAXSTACK = 10
    MAXLOCALS = 10
}
Loading _k_k_1_3_
defining _k_k_1_3_

+++ src/test/java/arb/expressions/nodes/nary/ProductTest.java
@@ -18,6 +18,6 @@ public class ProductTest extends
   public static void testOneTimesTwoTimesThreeEqualsSix() throws AnalyzerException
   {
     NullaryFunction<Real> prod = RealNullaryFunction.express("∏k{k=1…3}");
-    assertEquals(6.0, prod.evaluate(null, 128, new Real()));
+    assertEquals(6.0, prod.evaluate(null, 128, new Real()).doubleValue());
   }
 }
